removed max width and decoupled ActionBar component
