+++ packages/layouts-react/src/components/ActionBar/index.tsx
@@ -1,0 +1,2 @@
+export { ActionBar } from './ActionBar';
+export type { ActionBarProps } from './ActionBar';
