+++ packages/layouts-react/src/components/PageLayout/PageLayout.tsx
@@ -11,7 +11,6 @@ import { Header } from './Header';
 import SideNavigation from './SideNavigation';
 import TitleArea from './TitleArea';
 import BottomBar from './BottomBar';
-import ActionBar from './ActionBar';
 
 export type PageLayoutProps = {
   /**
@@ -35,6 +34,5 @@ PageLayout.Content = Content;
 PageLayout.ToolsArea = ToolsArea;
 PageLayout.TitleArea = TitleArea;
 PageLayout.BottomBar = BottomBar;
-PageLayout.ActionBar = ActionBar;
 
 export default PageLayout;
